third commit

+++ database/migrations/2026_01_20_071008_create_registered_users_table.php
@@ -17,7 +17,7 @@ return new class extends Migration
             $table->string('first_name');
             $table->string('last_name');
             $table->string('password');
-            $table->integer('user_type');
+            $table->unsignedBigInteger('user_type');
             $table->string('location');
             $table->string('created_by');
             $table->string('updated_by');
